requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pymupdf>=1.23.0
ocrmypdf>=14.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Web scraping
beautifulsoup4>=4.11.0
scrapy>=2.11.0
selenium>=4.15.0
playwright>=1.40.0
requests>=2.28.0
aiohttp>=3.9.0
lxml>=4.9.0

# Office documents
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
odfpy>=1.4.1
xlrd>=2.0.1

# eBooks
ebooklib>=0.18
mobi>=0.3.3
pypub>=1.5

# Text processing
nltk>=3.8.0
spacy>=3.7.0
langdetect>=1.0.9
ftfy>=6.1.0
chardet>=5.2.0
python-magic>=0.4.27

# Privacy and security
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
cryptography>=41.0.0

# Data formatting
datasets>=2.14.0
pyarrow>=14.0.0
fastparquet>=2023.10.0
jsonlines>=4.0.0

# Analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wordcloud>=1.9.0
scikit-learn>=1.3.0

# API and Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
websockets>=12.0
cors>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0

# Monitoring and logging
loguru>=0.7.0
prometheus-client>=0.19.0
sentry-sdk>=1.38.0

# Testing (optional, but included for completeness)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
faker>=20.0.0

# Development tools
black>=23.11.0
ruff>=0.1.0
mypy>=1.7.0
pre-commit>=3.5.0